requirements.txt
discord.py
openai
speechrecognition
pyaudio
gtts
yt-dlp
aiohttp
python-dotenv
PyNaCl